refactor: feedback follow up
